Move utils for backends out into the main backend module
